Standardized flowerbox width
Perhaps we should stop using flower boxes. But, for now, they continue to be part of our commenting standard. And if we're going to use them, they should consistently be 128 characters, per our file width standard. Most of them were. But there were a number of exceptions, likely replicated due to copying and pasting of old files (this code base is over a decade old, afterall).