Enforce file width standard
Ignia maintains a 128 character width standard for files. And, for the most part, that's honored. But a handful of areas—and especially comments, broke over the 128 character limit. In most case, this was likely due to refactoring or automated renames which made `<see/>` references longer. Regardless, the vast majority of these are fixed here. (There are a handful that were left as is due to Ignia not having a good multi-line standard for breaking up particular types of references; we'll want to establish a consistent way of handling those.)