Updated logic to rely on cached `@Topics` variable
Previously, every query included a subquery based on `RangeLeft` and `RangeRight` to ensure that only affected topic attributes were deleted. That's an expensive operation, and especially to repeat on multiple queries in a row. To mitigate this, I've stored the affected `TopicId`s in a `@Topics`  table variable, and done an `in` query against that. This should be much faster.