Wrapped `DeleteTopic` in transaction
Previously, any one query could fail and it would leave the database to an inconsistent state. Ideally, none of the queries should fail. But, if they do, we don't want to commit the transaction, but rather return an error. Now, we rollback a transaction if an error occurs, and then rethrow the original exception.