Wrapped `MoveTopic` in transaction
Technically, this may not be strictly necessary as the bulk of the logic is in a single (monster) query—and there's very little chace of the subsequent query failing. That said, it's still a best practice, and will also make it easier to restrict the scope of locks, should we add those in later (as is likely).