Introduced transaction isolation and locks
There are a lot of benefits to the nested set model used in `topics_Topics`, such as fast read performance. The tradeoff is that updates are incredibly invasive, and not friendly to concurrency. The hierarchy can become corrupted if multiple procedures attempt to modify the tree simultaneously. To help prevent this scenario, we've instituted `TABLOCK`s on `topics_Topics` during `topics_DeleteTopic` and `topics_MoveTopic`. Further, we've instituted the `SERIALIZABLE` transaction isolate level so that readers get a stable state during the entirety of the operation. (This also treats the `TABLOCK` as a `TABLOCKX, XHOLD`.) We don't use these lightly, as they greatly limit concurrency; in this case, however, that's a necessary, even if undesirable, precaution. Fortunately, write operations are comparatively rare—and, especially, move and delete operations.