Introduced new `AttributeValues` user defined table type
Previously, attribute values were passed to the `topics_CreateTopic` and `topics_UpdateTopic` stored procedures as lists concatenated with unique delimiters. This approach was ugly and potentially error prone. For instance, if the concactenated list were greater than `VARCHAR(MAX)`, it would get trunctated. Or, if an attribute value contained one of the delimiter characters, it would corrupt the data set. Both scenarios are highly unlikely. Nevertheless, user defined table types provide a more elegant solution by allowing the data to be passed as a semi-structured `DataTable` instead. On the stored procedure end, this allows us to simply treat the input as an existing table, instead of first parsing it using the custom `split()` function.