Broke `topics_MoveTopic` up into two `UPDATE` statements internally
The main body of the `topics_MoveTopic` was setup as one gigantic `UPDATE` query with multiple levels of nested `CASE` statements. This made it incredibly hard to read and intuit. The primary `CASE` statement had to do with whether the tree was moving left or right from its original location. That is an exclusive choice; as such, we can break this up into two distinct queries, while having confidence that only one of them will ever run. This is much, much easier to read, and it has no risk of impacting data performance or data integrity since it's the same logic as before, just refactored. In fact, initial intrumentation suggests it's probably a few miliseconds faster on average.