Introduced `WHERE` clauses on move `UPDATES`
The move updates have been incredibly slow. Much of this cost is due to updating the indexes. But the indexes are pretty basic, and efforts to modify them (e.g., moving the `CLUSTERED PRIMARY KEY` to a `UNIQUE INDEX`, or moving `RangeRight` to a _covered index_ using `INCLUDE()`) consistently yielded _worse_ results. That said, one major oversight is that the original queries didn't restrict the range of records impacted. Every update would modify EVERY `RangeLeft` and `RangeRight` value in the database, even if it was setting it to the original value. There were limits to the range, but only in the internal `CASE` statements, not in the `WHERE` clause. But limiting the scope of the `UPDATE` with `WHERE` statements, my test cases have gone down from a noisy range of 10-150ms, to a onsistent 3ms execution time. Other queries will take longer, and especially when they aren't working off of cached data. Nevertheless, the base query should be faster and, hopefully, this will yield much better results across the board.