Handled unused assignments
There are a couple of calls that are intended to warm the cache, and ensure that dependency libraries are in scope. These were previously assigned to variables. Those variables aren't used anywhere. A more efficient way of handling this which better communicates the purpose is to assign these to `_`. In addition, I preferred `var` over explicit setting of type, and removed an unused `_connectionString` field.