Optimized indexes on temporary table
The previous `#Topics` table had two indexes: The clustered index was on `SortOrder`, even though it's only used once, in one of the first queries. The second index was on `TopicID` _and_ `SortOrder`. This added overhead to filling the temporary table, but didn't offer much performance benefit in querying the data. To optimize this, I've replaced it with a single index on `TopicID`, which helps with the joins. This increases the cost of the initial `topics_TopicIndex` query a bit, but less than the cost of constructing the initial index. (Though, in practice, these are trivial gains.)