Centralized logic for version history
In a previous commit (#541a08d) we resolved a bug in how the `TOP 5` query was being constructed by refactoring the query to use a `GROUP BY` statement, thus allowing us to only query `UNIQUE` records. This update wasn't made to the `topics_GetVersion` query. It was also a fairly complex query, that didn't make sense to duplicate. For consistency, and to encapsulate the logic behind a simple object, we moved this to a new `topics_VersionHistoryIndex` view. It's worth noting that, unlike other queries in `topics_GetVersion`, we can call this from one of the centralize views because we don't actually version the queries returned; that's because a rollback will actually append the old data as a _new_ version, instead of doing a true rollback where interstitial versions are removed. That allows us to keep the queries simple in the sprocs, and centralize the complexity in the view.