Modified `Key` search to exact match
Previously, the `topics_GetTopicID` stored procedure used a `LIKE` query to search for a topic's `Key`. The fundamental idea of searching for a topic `Key` is a bit flawed since multiple topics can have the same key, and there isn't an (easy) way to search the database by unique keys. That said, if we _must_ allow this type of query, it should require an exact match, not a partial match.