Introduced script to detect potential sources of corruption
Hopefully, the nested set hierarchy remains in healthy shape—and with diligence when writing queries and maintaining appropriate locks during updates, it should. However, because the nested set can be susceptible to corruption, it's important to have a way to detect and isolate problems, should they occur. This stored procedure is incredibly slow (there's no optimization for speed, and the queries are pretty ugly), but it will detect the three scenarios that indicate data corruption and help isolate them down to affected topics. That said, it makes no attempt to fix the issues; that will always require manual intervention to review the specific cases and determine what should be done.