Ensure first topic in hierarchy is selected
`topics_GetTopicID` artificially limits the results using a `TOP 1`. There may be multiple topics for a given `@TopicKey`. Given no better method for determining which one the user _intended_, we should choose the one that's first in the _hieararchy_ (whereas previously, it chose the one that had been _created_ first). I don't know than any of these are optimal; it'd be better to get be able to query by fully-qualified path. That said, in almost all scenarios, we're just using this to get the root node. For that reason, I also added in an optimization hint indicating that.