Implemented newly optimized `topics_GetTopicID`
`topics_GetTopics` was essentially duplicating the same logic from `topics_GetTopicID`. To simplify and centralize the code, we're using the stored procedure instead. This also fixes an issue with the original default, which would select the first `TopicId` assuming it was the root; that's _generally_ the case, but it's not _guaranteed_. (E.g., topics could be moved to a new root later, theoretically.)