Removed `topics_` prefix
The `topics_` prefix is an Ignia standard introduced based on the assumption that multiple applications may be living in the same database. This was inherited from legacy corporate rules where tenant sites were provisioned a single database. In practice, today, the **OnTopic** database is exclusively placed in its own dedicated database—which is a configuration we recommend. Given this, the use of the database prefixes is unnecessary and cumbersome. While I was at it, I also moved all objects out of the `dbo` folder since we have no intention of introducing additional schemas to the database.