Renamed `TopicAttributes` to `Attributes`
The `topics_TopicAttributes` name was a legacy of the fact that prior to introducing the Oroborous Configuration, we had a `topics_Attributes` table with a schema for defining attributes. Since those attributes are now defined as standard topics, that table is no longer used. Since it's been long deprecated, and we're renaming tables anyway, this is a good opportunity to give it a more consistent name.