Renamed `Blob` to `Blobs`
Our conversion is to pluralize tables, but the `Blob` table was always singular. Technically, we should probably revist the name in general—the ancillary attributes are no longer stored as blobs, but rather native XML, since SQL now has a built in data type for that. But we'll need to think of a better name for it before making that change.