Rename `Blobs` to `ExtendedAttributes`
The name `Blobs` has always been a bit of a misnomer, originating from a very early version of **OnTopic** which stored extende attributes in a binary format. Since they are now (and have been for a long while) stored using SQL's built-in XML data type, a different name makes sense. I've renamed these to `ExtendedAttributes` with the actual `Blob` field to `AttributesXml` (since column names should be distinct from table names, another problem with the legacy implementation). As part of this, I also renamed the `StoreInBlob` property to `IsExtendedAttribute`, which is a breaking change on the topic library. This also required a lot of updates to comments, which made extensive reference to the blob storage.