Added validation check for mismatched parents
The `ParentId` is a cached version of the nested set model's parent topic. If these two values don't match, that's a sign of corruption. Usually, this means that an interstitial topic was deleted and no longer exists, but its children failed to get deleted. In that case, the hierarchy is actually intact, but won't get reconstructed correctly based on the `ParentID` value.