Exclude `ParentID` from update statements
If the `ParentID` has changed, `Move` should be called. This logic is enforced by the `BaseTopicRepository.Save()` method, which explicitly checks for changes in the `Parent.Id`. That said, even in this scenario, the `ParentID` should not be updated. There are only two places it should be set: `CreateTopic` and `MoveTopic`. The latter will always update the initial version, preventing parenting. While this prevents us from versioning or rolling back the hierarchy, that exact _also_ prevents a lot of potential issues like rolling back the hierarchy to a parent that no longer exists, or to a key that now causes a conflict. Generally, we see this solving more problems than it creates. This has long been the logic, but because `ParentID` might get rewritten in `UpdateTopic`, this would yield duplicate `ParentID`s. (The update in `MoveTopic` doesn't limit by `Version`, and so will update all versions in the case of a move.)