Removed `IsDraft` placeholder
At one point, there was talk of introducing `IsDraft` functionality so that drafts could be created and previewed without going live. This hasn't proven critical since new topics can be created as `IsHidden`, and it really only works if there's a separate cache/server for drafting/previewing topics. For that reason, it was never fully implemented. We shouldn't confuse the interface with partially implemented functionality, so we're removing it. If we choose to revisit this feature in the future, we can add it back in then.