Removed `@ParentID` from `UpdateTopic`
`UpdateTopic` has no need for the `ParentId` since this should only be updated as part of `MoveTopic`. As such, to help prevent ambiguity over functionality and remove unnecessary parameters, this is being removed. It is still needed, obviously, for `CreateTopic`, however, so the call in `SqlTopicRepository` is now conditional, only being evaluated if both the `Parent` is not null (preexisting condition) _and_ the `Topic.Id` is -1 (meaning, it's a new topic).