Override `ModelType`, `IsExtendedAttribute`, and `EditorType` as appropriate
In previous versions of the editor, the editor type was determined by a `FileList` attribute mapping to **ASP.NET WebForms User Controls**, and the `ModelType` was conditionally set based on the name of that file. As of v4.0.0+, however, the editor type is determined by the content type, which maps to a concrete class in the `AttributeTypes` namespace. This allows us to set the `EditorType` based on the class name, and override the `ModelType` on a per class basis depending on how the data is modeled. Further, since many attribute types correspond to an expected length, we can override `IsExtendedAttribute` as appropriate (e.g., setting it to `true` for `HtmlAttribute`. To facilitate all of this, a new `AttributeTypeDescriptor` base class was introduced to set defaults for eachof the attribute types—while the underlying `AttributeDescriptor` maintains logic that is appropriate for the legacy approach. That way, older versions of the editor, which rely exclusively on `AttributeDescriptor`, will inherit that logic, while updated versions will inherit from `AttributeTypeDescriptor`'s overrides, if not type-specific overrides in concrete classes.