Migrated from *.NET Framework 4.8* to *.NET Standard 2.0*
Previously, the core libraries were dual targeting **.NET Framework 4.8** (`net48`) and **.NET Standard 2.1** (`netstandard2.1`). This works fine. But since these are fundamentally .NET Standard classes, and .NET Framework can utilize .NET Standard classes, it makes sense to use that across the board. That said, the latest version of .NET Standard that .NET Framework (4.3.1+) can utilize is .NET Standard 2.0, so at minimum we need to support that for backward compatibility with `Ignia.Topics.Web` and `Ignia.Topics.Web.Mvc`. Technically, we don't _need_ to support .NET Standard 3.1, but doing so allows us to use the OOTB implementation of nullability hints, such as `[AllowNulls]`, which are otherwise dependent on polyfills in previous versions.