Only insert `ExtendedAttributes` if they've changed
Since we introduced versioning, we've only inserted `Attributes` if they differ from the last version. That logic wasn't put in place for `ExtendedAttributes`, however, since we assumed that most updates _would_ change them (e.g., would be to the `Title` or `Body`). In practice, that hasn't necessarily been the case (e.g., perhaps because of changes to visibility, URL, relationships, &c.) As a result, there are a lot of unnecessary duplicates stored in the `ExtendedAttributes` table. That stops here, as we'll only write the version if it is distinct from the previous version. (In the future, we should wrote a script to get rid of all previous duplicates.)