Moved strongly typed `Get{Type}Value()` and `Set{Type}Value()` to extensions
The `AttributeValueCollection` has a series of strongly typed overloads to `GetValue()` and `SetValue()` that allow specific types (such as `Boolean`, `Integer`, &c.) to be retrieved or set from the attributes, thus taking care of any conversion necessary. These are ancillary requirements that not everyone will require, however, and take up a lot of space. As such, they're being migrated to an extension method in the `Attributes` namespace. As part of this, we're also moving `[AttributeSetter]` and `AttributeValue` to the new `Attributes`