Validate version in `Load()`
Normally, when we call `Load()`, we expect a `null` response if a specified topic cannot be found. With `Load(version)`, however, an exception makes more sense, as this should only be called when a specific version is _known_ to exist (e.g., as part of a call to `Rollback()`). Further, this provides better error handling as otherwise an exception is thrown on `SetVersionHistory()` since a valid topic with an invalid version won't return any topic records, but will return a version history (since the version history isn't version specific). That was previously throwing an unintuitive exception regarding the `topicId` not being in the `topics` collection. This is a better way of handling it.