Introduce basic concurrency check
Normally, we'd avoid the double-checked locking (anti-?)pattern, but in this case it should be alright since we never modify the collection outside of this method. Given that, one might argue that we shouldn't expose the collection publicly, and I'd agree. That's going to require more consideration, though, before implementing. For now, as a class in the `Internal` namespace, it's not a top priority.