Implemented `<inheritdoc />`
Previously, we needed to repeat documentation on implementations of interfaces or derived types. This was repetitive, error prone, and bloated the code. In **Visual Studio 2019**, we finally have tooling for `<inheritdoc />`—which is already supported by many XML Doc processors, like **SandCastle**. This allows us to inherit documentation from parent classes or interfaces, and thus only maintain one set of code.