Marked `DefaultConfiguration` members as `[Obsolete]`
The `DefaultConfiguration` (and related `Configuration` and `GetConfigurationValue()` members) is a left-over of the legacy `AttributeDescriptor` model where attribute type configuration values were configured as a delimited string in an attribute. This is still required for backward compatibility with the legacy **ASP.NET WebForms** editor. The new **ASP.NET Core** editor does not use this, instead relying on the new `AttributeTypeDescriptor` classes, which provide direct access to each configuration value via the standard `Attributes.GetValue()` approach.