Updated `StubTopicRepository` to use new `AttributeDescriptor` format
To more accurately test the current assumptions, updated the `StubTopicRepository` to create the new `AttributeDescriptor` format. This necessitated creating a couple of basic view models to accommodate mapping of a `TextAttribute` and `TopicReferenceAttribute`. It also required updating a few tests to accommodate the change in `ContentTypeDescriptor`s count, and the new `EditorType` names (e.g., `TextAttribute` instead of `FormField`).