Documented the `ViewModels`
Typically, we don't document view models, since they're meant to be quick and easy to program, and usually the properties are self-explanatory. That said, since the `Ignia.Topics.ViewModels` collection represents a fairly stable and semi-permanent set of view models, which will be used by most implementations of **OnTopic**, it makes sense to provide documentation, not only for any generated reference documentation we may produce in the future, but also to provide an aid in IDEs (e.g., for _Intellisense_, for users of **Visual Studio**).