Introduced `LayoutController` and affiliated views
In the **ASP.NET MVC Framework**, we use an embedded `LayoutController` instead of the newer view component functionality. It's effectively the same technique, just not quite as optimized for individual components. That said, the views are identical—outside of the fact that we can use an actual HTML `@Helper` in this version.