Introduced views into `Ignia.Topics.Web.Host`
There's actually only one view I'm bothering to create, since this is already an exercise in formality, and that's for the most ubiquitous `Page.aspx`. Other views are easy to produce based off the template, should anyone ever have the desire. That said, we should really be putting the `Ignia.Topics.Web` projec to pasture at this point.