Establish new release branch for 4.0.0
This branch is being established as an `--orphan` so that we can
perform an explicit merge on develop, effectively hiding all
previous commits from e.g. `git log --first-parent`. In the
future, this will allow us to utilize git log `--first-parent`
as a quick way of rolling up all merges on integration branches—
such as `develop` and `master`—without cluttering the history
with all of the commits that went into e.g. feature branches.
That said, the commits made on contributing branches will still
be available via e.g. `git log` or `git blame`, so line-by-line
modifications can be tracked back to specific commits, instead
of squashed into a single merge.

For more  detail see the following StackOverflow question:
https://stackoverflow.com/questions/59507140/establishing-a-
clean-git-history-for-explicit-merges/59507446