Updated documentation for 4.0.0 release
Did an initial review of the documentation to account for the 4.0.0 release. Most importantly, updated the references from `Ignia.Topics` to `OnTopic`, as well as the references to `OnTopic.Web` and `OnTopic.Web.Mvc` to point to their new external repositories. In addition, updated the language as appropriate while reviewing the documentation.