Introduced `Relation` and `RelationWithChildren` view models
The `RelationTopicViewModel` and the `RelationWithChildrenTopicViewModel` specifically service the `TopicMappingServiceTest.RecursiveRelationships()` test by providing a bare bones DTO for testing cousins, second cousins, and cousins once removed. As part of this, I made minor adjustments to `RecursiveRelationships()` to clarify the relationships of each object, and to introduce local helper functions. Also applied this to `MapRelationships()`.