Adjusted `Descendent` view model to return a `TopicViewModelCollection`
While this necessitates deriving `DescendentTopicViewModel` from `TopicViewModel`—and, thus, mapping a number of other fields—it also allows us to test the `FilterByContentType()` test `GetByContentType()` using the same (virtually identital) view model. While we're not explicitly testing the properties inherited by `TopicViewModel`, a base of `TopicViewModel` is necessary to construct a `TopicViewModelCollection`. In practice, this is the collection many view models will use.