Introduced shared `_mappingService` with `DummyTopicRepository`
Only a few tests actually depend on a semi-functional `ITopicRepository` (i.e., by looking up data in a `StubTopicRepository` for e.g. a mapped `TopicReference` or `[Metadata()]`). For the rest, they can be simplified by using a shared `_mappingService` which uses a `DummyTopicRepository` with no functional components. This reduces the number of times we need to instantiate a new `TopicMappingService` or `FakeViewModelLookupService`, and makes it clearer when we're genuinely relying on a test which is actually dependent on the `ITopicRepository`, since most mapping operations are not.