Introduce `GetChildTopic()` helper functions
Quite a few of the tests rely on checking a collection to determine if a view model with a particular key exists. The code for this is long, repetitive, and clunky. In recent refactorings (see e.g. commits references from #9225f23), quite a few of these were replaced with local helper functions (e.g., `getChildTopic()`, `getRelation()`, &c.). That said, nearly all instances of these can be handled by two overloads of a function (one for view models deriving from `KeyOnlyTopicViewModel`, and another for those deriving from `TopicViewModel`). As such, this refactoring replaces nearly every call that depends on verifying that a child topic exists with a new `GetChildTopic()` helper function, and removes any redundant local functions that served the same purpose.