Removed unused and unnecessary `Configuration` property
The `Configuration` property dictionary previously mapped to the `AttributeDescriptor.Configuration` property, which was removed as part of the 4.0.0 migration to `AttributeTypeDescriptor` based configuration. As such, it no longer provides a reliable test for evaluating compatible type property mapping. That said, we are effectively testing that with `VersionHistory` and `ModelType`, so it's not strictly necessary, either.