Preferred generic `MapAsync<>()` in favor of casting to a view model
In terms of the test case, this generally yields the same outcome. It's a bit faster, however, as it doesn't need to rely on the `ITypeLookupService` to determine the appropriate topic to create, nor reflection to create that topic. I also find the syntax just a tiny bit more readible—though that's a pretty nit-picky difference here. The one obvious exception is `MapDynamic()` since it is explicitly intended to test the ability to dynamically determine the correct view model based on the view model name.