Corrected bug with `Contract.Assume(false)`
Previously, `Contract.Assume(false)` was _always_ evaluating as _true_, and thus not returning an exception. This is because while there was an overload for `Contract.Assume<T>(bool)` there was not one for `Contract.Assume(bool)`, and so it was getting handled by the `Contract.Assume(object)` overload. And since `false` is not a null object, it wasn't throwing an exceptions. Whoops! A win for expanded test coverage (see #f5cbd39).