Implemented more expressive test names
For consistency with e.g. `AttributeValueCollectionTest` and `ITopicRepositoryTest`, updated the test namesin `RelatedTopicCollectionTest` to be more expressive. Because of the nature of the current tests, the full format of `METHOD_CONDITION_RESULT` doesn't really make sense, as most of these tests don't vary by condition, and so I've shortened it to `METHOD_RESULT`. We may revisit this later if we expand the test coverage.