Incorporated contextual `attributePrefix` into `PropertyConfiguration`
Previously, `PropertyConfiguration` was aware of the `AttributePrefix` defined on the local property via the `[MapToParent]` attribute, but it wasn't aware of any _contextual_ `attributePrefix` parameters that the property might be inheriting (i.e., from a parent property). The consequence of this is that the `AttributeKey` didn't necessarily reflect the _actual_ attribute key, and thus the `attributePrefix` needed to be relayed and prepended anywhere that `AttributeKey` was called. This is silly. The updated model allows the `attributePrefix` to optionally be sent to the `PropertyConfiguration` constructor and, if present, it is automatically prepended to both the `AttributePrefix` (to faciliate deep models) as well as the `AttributeKey` (so that the `AttributeKey` always reflects the actual key of the target attribute). Because the `attributePrefix` parameter defaults to empty, this won't affect any prefix implementations until they start passing the `attributePrefix` parameter.