Retrofitted `ReverseTopicMappingService` to use new `PropertyConfiguration`
In #8bb6ab4, `PropertyConfiguration` was updated to be made aware of the contextual `AttributePrefix`, thus allowing the `attributePrefix` to be relayed (and composited) via the `PropertyConfiguration` instead of by passing it to each helper method, and then prepending it to every call to `configuration.AttributeKey`. This is much more elegant, and more intuitive from an implementation perspective. This doesn't change the functionality, just refactors the implementation so that the `attributePrefix` is more deeply integrated into the infrastructure—which, conveniently, makes it so that most downstream elements no longer need to be aware of the `attributePrefix`. As an example of this, we were able to remove `attributePrefix` entirely from the `BindingModelValidator` since it now infers it via the `PropertyConfiguration.AttributeKey`.