Introduced `[MapToParent]` functionality to `TopicMappingService`
The `[MapToParent]` capability was originally introduced as part of the `ReverseTopicMappingService`, to allow complex binding models to be mapped back to a single `Topic` entity. There's no reason the opposite shouldn't also be supported—nor is it all that difficult to add this feature. This commit introduces the initial versionof this.