Return `Version` for `Attributes` and `ExtendedAttributes`
Previously, `Version` was not returned for _any_ fields, only as a standalone value for the topic. In preparation for supporting import scenarios where having a last updated date is important, we are returning `Version` with `GetTopics`—and, thus, the `AttributeIndex` and `ExtendedAttributeIndex` views it depends upon. We will not be returning `Version` for key fields (i.e., `Key`, `ContentType`, and `ParentId`).