Added (optional) `lastUpdated` parameter to constructor
Previously, the `AttributeValue` property had a `LastUpdated` property, but as it was automatically assigned, it wasn't included in the constructor. In preparation for allowing this to be set as part of the loading and import processes, however, this needs to be exposed as a writeable field, at least via the constructor (since, currently, the property itself is readonly). Arguably, this _might_ make more sense being placed after `value`; since mapped properties will be calling `enforceBusinessLogic`, however, but _not_ supplying `lastModified`, we're placing it last. Regardless, as ancillary metdata on an internal overload, the order isn't too critical here.