Allow `LastModified` to be (internally) updated
Previously, `LastModified` was treated as readonly, to align with the original expectation that `AttributeValue` would be immutable. That's not how it's ended up being treated, however, and here we have a scenario where the `LastModified` needs to be updated. This is because `EnforceBusinessLogic` will trigger `SetValue()` indirectly via an `[AttributeSetter]` property. That's fine for updating the value. But for load and import scenarios, the `[AttributeSetter]` property will orphan the `Version` date intended to be used for the `LastModified` property. As such, we're exposing a setter to allow that property to be updated after this occurs. Since this is an outside scenario intended only to support infrastructure, however, the setter is marked as `internal`; there is no supported scenario where this should be updated directly by external callers.