Added `Version` field to `SetValue()`
In order to allow the `AttributeValue` to be annotated with `LastModified` metadata, we're adding `Version` to the `SetValue()` calls, based on data retrieved from the `GetTopics` stored procedure. Because the published schema for 4.0.0 doesn't return a `Version` field when calling `GetTopics`, this code checks for the number of returned columns to maintain backward compatibility; this allows this to be published as an minor version without breaking semantic versioning rules. Preferably, we'd actually check for the existence of the `value` property, but that's quite a bit more involved with `SqlDataReader`, and the column count gives us confidence in that, at least for this version. (We'll ideally want to remove this in the 5.0.0 release.)