Updated test to properly evaluate sitemap
The previous test was designed to fail—and expect an exception. That created a false sense of comfort, as the core functionality wasn't actually being tested. This issue is resolved, and the test has been updated to look for `utf-8` encoding, along with a `standalone` attribute. The former is required by sitemap.org's schema. The latter will be written by `XDocument` serialization, assuming we continue to use `XDeclaration`.