Changed how encoding is handled
Previously, to ensure that the XML delaration was being written, a `StringWriter` was used. `StringWriter` _always_ outputs `UTF-16`, however, regardless of what the `XDeclaration` is set to. That's a problem, as Sitemap.org _mandates_ `UTF-8` and e.g. Google will reject a sitemap that uses `UTF-16`. To mitigate this, the `StringWriter` is remoted and, instead, a simple approach is taken: simply declare a standalone `XElement` and serialize it to a string alongside the `XDocument`. (Note: Calling `ToString()` on an `XDocument` won't include the `XDeclaration`, which is why the `StringWriter` had previously been used.)