Ensure nullable `RangeRight` column uses `ISNULL()` fallback
Per `SR0007` warnings, nullable columns can bubble up in the results, yielding nullable predicates. In practice, `RangeRight` is only nullable for import scenarios, and is expected to have a value. That said, to be certain, we're using `ISNULL(RangeRight, -1)` anytime it's being referenced. `-1` is used since the `RangeLeft` to `RangeRight` should always be positive; this helps avoid scenario where a `null` triggers a false positive. Again, though, this _shouldn't_ happen in a real-world scenario.