Return explicit column list
While it shouldn't hurt anything here since `ValidateHierarchy` is not expected to be called by code, but rather by humans, we should maintain best practices by returning an explicit list of columns instead of relying on wildcards. Otherwise, if the column list or order changes in the table, the output of the procedure will change. Explicitly listing the columns (and order) helps insulate downstream consumers from such potential changes in the future.