Explicitly cast `TopicID`
When doing a comparison, the `AttributeValue` was implicitly being cast from `VARCHAR(255)` to `INT`. In our scenarios, this shouldn't be a problem. But because it could result in data loss if the `ParentID` field is not (correctly) set to an integer, that raises a warning. To avoid this—and to provide a very trivial performance improvement—we're explicitly converting the `TopicID` from `INT` to `VARCHAR(255)`. This yields no data loss, and allows the two fields to be quickly compared to one another.