Update `GitVersionTask` to latest version
We are currently using one of the early releases of GitVersion 5.x. There have since been quite a few updates, with the latest being `5.2.4`. That said, the `5.2.x` versions seem to have a number of problems, with multiple patches having been released over the last two weeks. Further, the latest yields errors with the `GitVersionTask.targets`. Given that, we're going to stick with the `5.1.x` release from December, which appears to be stable; we'll reevaluate the `5.2.x` version in a few weeks, once it has (hopefully) stabilized. Fortunately, as a development-only dependency, this doesn't affect downstream consumers, just our internal build process.