Update `GitVersionTask` to latest version
Upgraded to the 5.2.x release of `GitVersionTask`. As an internal (private) dependency only used by our build process, this shouldn't affect downstream consumers and, thus, can be kept on the latest version. That said, we had previously kept with the 5.1.x release due to what appeared to be some churn and instability in the 5.2.x release (see #9348b9c). Now that the branch has stabilized, and a confirmed fix for a bug in Azure DevOps has been posted, we should be able to utilize the latest version.